インターフェースのアドレス管理の実装途中

+++ sys/kern/pktbuf.c
@@ -40,7 +40,7 @@ void pktbuf_remove_header(struct pktbuf *pkt, size_t size) {
   pkt->head += size;
 }
 
-void pktbuf_copyin(struct pktbuf *pkt, u8 *data, size_t size, off_t offset) {
+void pktbuf_copyin(struct pktbuf *pkt, u8 *data, size_t size, u32 offset) {
   memcpy(pkt->head+offset, data, size);
 }
